requirements.txt
streamlit>=1.28,<3
pandas>=2.0,<3
pyarrow>=14.0,<19
plotly>=5.18,<7
altair>=5.0,<6
openai>=1.10,<3
python-dotenv>=1.0,<2
faker>=22.0,<35
